+++ spotifywebapipython/zeroconfapi/__init__.py
@@ -1,0 +1,21 @@
+# import all classes from the namespace.
+from .blobbuilder import BlobBuilder
+from .credentials import Credentials
+from .spotifyzeroconfapierror import SpotifyZeroconfApiError
+from .zeroconfconnect import ZeroconfConnect
+from .zeroconfresponse import ZeroconfResponse
+from .zeroconfgetinfo import ZeroconfGetInfo
+from .zeroconfgetinfoalias import ZeroconfGetInfoAlias
+from .zeroconfgetinfodrmmediaformat import ZeroconfGetInfoDrmMediaFormat
+
+# all classes to import when "import *" is specified.
+__all__ = [
+    'BlobBuilder', 
+    'Credentials',
+    'SpotifyZeroconfApiError',
+    'ZeroconfConnect',
+    'ZeroconfGetInfo',
+    'ZeroconfGetInfoAlias',
+    'ZeroconfGetInfoDrmMediaFormat',
+    'ZeroconfResponse',
+]